moved blending logic to cupBlender.js
