+++ lib/pourer.js
@@ -29,6 +29,8 @@ var pourer = (function () {
             internalPour(testName, testFn, TEST_TYPES.WRAP);
         };
 
+        debug("attaching pour to cup: " + cup.name);
+
         cup.pour = internalPour;
     }
