+++ lib/stirrer.js
@@ -1,7 +1,7 @@
 var sinon = require("sinon"),
     debug = require("debug")("mocha-stirrer:stirrer"),
-    pourer = require("./pourer"),
     Mocker = require("./RequireMocker"),
+    cupBlender = require("./cupBlender"),
     utils = require("./utils");
 
 //todo: implement clear method to remove befores/afters/pars/fakes
@@ -51,7 +51,7 @@ var stirrer = (function () {
 
         conf = conf || {};
 
-        var cup = _blend(conf); //create the Cup with sinon sandbox and all the goodness of stirrer
+        var cup = cupBlender.blend(conf); //create the Cup with sinon sandbox and all the goodness of stirrer
 
         if (conf.setupImmediate) { //if told to, setup everything right away, if not then setup will happen in the before hook
             _setup(cup, conf);
@@ -67,144 +67,14 @@ var stirrer = (function () {
     }
 
     function restir(cup) {
-        cup.sb.restore();
+        cupBlender.restir(cup);
     }
 
     function fakeRequire(cup, requirePath, options) {
-
-        var reqModule = cup._mocker.require(module.parent, requirePath, options);
-
-        _enrichCupStubsFromMocker(cup);
-
-        return reqModule;
+        cupBlender.fakeRequire(module.parent, cup, requirePath, options);
     }
 
 //******************* PRIVATE METHODS *****************
-
-    var cupBody = (function () {
-
-        /**
-         *
-         * @param befores is either a function or an array of functions
-         */
-        function _stirBefores(befores) {
-
-            var stirFn = function () {
-                debug("stirring in the registered befores - " + this.name);
-                if (utils.isFunc(befores)) {
-                    befores = [befores];
-                }
-
-                befores.forEach(function (fn) {   //todo: use concat
-                    this._befores.push(fn);
-                }.bind(this));
-            }.bind(this);
-
-            if (befores) {
-                before(stirFn);
-            }
-        }
-
-        /**
-         *
-         * @param afters is either a function or an array of functions
-         */
-        function _stirAfters(afters) {
-
-            var stirFn = function () {
-                debug("stirring in the registered afters - " + this.name);
-                if (utils.isFunc(afters)) {
-                    afters = [afters];
-                }
-
-                afters.forEach(function (fn) { //todo: use concat
-                    this._afters.push(fn);
-                }.bind(this));
-            }.bind(this);
-
-            if (afters) {
-                before(stirFn);
-            }
-        }
-
-        function _stirPars(pars) {
-
-            if (pars) {
-                before(function () {
-                    debug("stirring in the registered pars - " + this.name);
-                    pars = utils.isFunc(pars) ? pars() : pars;
-                    this.pars = utils.merge(pars, this.pars); //add new pars to the pars map
-                }.bind(this));
-            }
-        }
-
-        function _stir(conf) {
-            _stirPars.call(this, conf.pars);
-            _stirBefores.call(this, conf.befores);
-            _stirAfters.call(this, conf.afters);
-        }
-
-        return {
-            require: function (path, options) {
-                debug("cup require called with path = " + path + " - " + this.name);
-                return fakeRequire(this, path, options);
-            },
-
-            stir: function (conf) {
-                debug("cup stir called - " + this.name);
-                _stir.call(this, conf);
-            },
-
-            restir: function () {
-                debug("cup restir called - " + this.name);
-                restir(this);
-            }
-        };
-    })();
-
-    function _blend(conf) {
-
-        var name = conf.name || Date.now();
-        var sbConf = conf.sandbox || void 0;
-        var sb = sinon.sandbox.create(sbConf);
-
-        if (!conf.delay) {
-            _setHooks(conf);
-        }
-
-        var cup = {
-            name: name,
-            sb: sb,
-            pars: {},
-            _mocker: new Mocker(sb),
-            _befores: [],
-            _afters: []
-        };
-
-        utils.merge(cupBody, cup);
-        pourer.attach(cup);
-        cup.stir(conf);
-
-        return cup;
-    }
-
-    function _setHooks(conf) {
-        if (!utils.isFunc(conf.before)) {
-            conf.before = _noop;
-        }
-
-        if (!utils.isFunc(conf.after)) {
-            conf.after = _noop;
-        }
-
-        if (!utils.isFunc(conf.beforeEach)) {
-            conf.beforeEach = _noop;
-        }
-
-        if (!utils.isFunc(conf.afterEach)) {
-            conf.afterEach = _noop;
-        }
-    }
 
     function _doHooks(cup, conf) {
 
@@ -404,13 +274,6 @@ var stirrer = (function () {
                 }
             }
         });
-    }
-
-    function _enrichCupStubsFromMocker(cup) {
-        //todo: implement !!!!
-    }
-
-    function _noop() {
     }
 
     return {
