+++ lib/utils.js
@@ -1,4 +1,4 @@
-module.exports = (function () {
+var utils = (function () {
     "use strict";
 
     function isFunc(fn) {
@@ -132,3 +132,5 @@ module.exports = (function () {
     };
 })();
 
+//******************** EXPORT ********************
+module.exports = utils;
